+++ util/zs_vector.h
@@ -6,14 +6,14 @@ namespace zetscript{
 	public:
 		//public
 		intptr_t* items;
-		uint16_t count; //number of items
+		unsigned int count; //number of items
 
 		//ZPList  	*	ZPList_New(void);
 		zs_vector();
 
-		void 		set( uint16_t idx, intptr_t e);
-		intptr_t	get( uint16_t idx);
-		void 		erase( uint16_t idx);
+		void 		set( unsigned int  idx, intptr_t e);
+		intptr_t	get( unsigned int  idx);
+		void 		erase( unsigned int  idx);
 		/**
 		 * Insert item at the end
 		 */
@@ -22,7 +22,7 @@ namespace zetscript{
 		/**
 		 * Insert item at position idx.
 		 */
-		void 		insert(uint16_t idx, intptr_t e);
+		void 		insert(unsigned int  idx, intptr_t e);
 
 		void 		clear();
 		intptr_t    pop_back();
@@ -34,7 +34,7 @@ namespace zetscript{
 		~zs_vector();
 	private:
 		//private
-		uint16_t _size; // size vector (user count for iterate through items)
+		unsigned int _size; // size vector (user count for iterate through items)
 		bool	push_back_slot();
 
 	};
